Set YOTTA_CFG_UVISOR_PRESENT to 0 (uVisor unsupported)
The YOTTA_CFG_UVISOR_PRESENT symbol should be 0 until the uVisor
is not officially supported for this target.

Having the config set to 1 triggers some overrides and API calls
that are not supported if uVisor is not present, causing either
compile-time or run-time errors.

If you are in the process of porting or experimenting uVisor on your
target, you should consider using an app-specific yotta config:
yotta --config='{"uvisor":{"present": 1}}' build
so that the setting does not affect the default value of the public
module.